This commit was manufactured by cvs2svn to create tag 'wxPy_2_4_0_4'.

+++ include/wx/dc.h
@@ -602,7 +602,7 @@ public:
             float *descent = NULL, float *externalLeading = NULL,
             wxFont *theFont = NULL, bool use16bit = FALSE) const ;
     void GetSize(float* width, float* height) const { int w, h; GetSize(& w, & h); *width = w; *height = h; }
-    void GetSizeMM(float *width, float *height) const { int w, h; GetSizeMM(& w, & h); *width = (float) w; *height = (float) h; }
+    void GetSizeMM(float *width, float *height) const { long w, h; GetSizeMM(& w, & h); *width = (float) w; *height = (float) h; }
 
 #endif // WXWIN_COMPATIBILITY
 
@@ -770,6 +770,8 @@ private:
     #include "wx/mac/dc.h"
 #elif defined(__WXPM__)
     #include "wx/os2/dc.h"
+#elif defined(__WXSTUBS__)
+    #include "wx/stubs/dc.h"
 #endif
 
 // ----------------------------------------------------------------------------
